pyproject.toml
[project]
name = "langchain-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.44",
    "langchain>=0.3.20",
    "langchain-google-genai>=2.1.0",
    "langchain-openai>=0.3.8",
    "python-dotenv>=1.0.1",
    "langchain-mistralai>=0.2.7",
    "langchain-google-firestore>=0.5.0",
    "chromadb>=0.6.3",
    "langchain-chroma>=0.2.2",
    "langchainhub>=0.1.21",
]
